src/test/java/med/voll/api/domain/endereco/ConsultaEnderecoServiceTest.java
package med.voll.api.domain.endereco;

import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
import med.voll.api.domain.endereco.dto.ViaCepResponse;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.RegisterExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.DynamicPropertyRegistry;
import org.springframework.test.context.DynamicPropertySource;
import org.springframework.web.client.RestTemplate;

import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

@SpringBootTest
@ActiveProfiles("test")
public class ConsultaEnderecoServiceTest {

    @RegisterExtension
    static WireMockExtension wireMock = WireMockExtension.newInstance()
            .options(wireMockConfig()
                    .dynamicPort()
                    .usingFilesUnderClasspath("wiremock"))
            .build();


    @Autowired
    private ConsultaEnderecoService consultaEnderecoService;

    @DynamicPropertySource
    static void registerWireMockProperties(DynamicPropertyRegistry registry) {
        registry.add("viacep.server", wireMock::baseUrl);
    }


    @Test
    void consultarEnderecoPorCepReturnsValidEndereco() {
        Endereco endereco = consultaEnderecoService.consultarEnderecoPorCep("01001-000");

        // Gere assertivas para verificar se o endereço está correto
        assertNotNull(endereco);
        assertEquals("Praça da Sé", endereco.getLogradouro());
        assertEquals("Sé", endereco.getBairro());
        assertEquals("01001-000", endereco.getCep());
        assertEquals("São Paulo", endereco.getCidade());
        assertEquals("SP", endereco.getUf());
        assertEquals("lado ímpar", endereco.getComplemento());


    }

}